lt-qcom-ubuntu-images-3.4: add new build job using vendor 3.4 kernel
Until all features available in the downstream 3.4 kernel are available in the
LT branch, let's keep a live build with 3.4. Release notes will give
instructions on how to install the 3.4 kernel .deb and boot 3.4 as an alternate
kernel.

Build only the -developer image, since we mainly need the boot.img and the
kernel .deb file after all. We use the -developer image for basic CI loop, though.

Change-Id: I6fbd0a466f5cd415b349f585e6297ad068c58593
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>