lt-qcom-ubuntu-images-arm64: add support for Dragonboard 410c
Add initial support for Qualcomm APQ8016 SBC Evaluation board.
This board is also referred to as the DragonBoard 410c.

Since the boot image now includes more than 1 DTB, also rename the boot image
file to a generic name, as it is expected to be used for all 64bit platforms
from QCOM, instead of 1 boot file per board as it used to be on armv7.

Change-Id: Ie0645fb0906b2955939128d9bd245c5d99423dab
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>