generic-package-builder: job for building debian packages
As seems often needed, a job for building a debian package
for linaro builds. Because of issues, implemented as matrix
job that builds unconditionally on all archs. If some packages
are only intended for selected architectures, just ignore failures.

Notice the slave list and post-build-repo list needs to be kept
in sync manually.

wishlist: groovy script that sends failure email not to me but:

1) if job was manually submitted, to the user who submitted the job
2) if triggered from upstream job, to the email list of the upstream job
3) the maintainer field in changelog if it ends with @linaro.org

Change-Id: I190bb3e8066945383263586ccda97716d168d606
Signed-off-by: Riku Voipio <riku.voipio@linaro.org>