khilman-trigger-flex: drop "-test" suffix from branch name when comparing
When determining whether a given branch needs to be triggered, drop
the "-test" suffix.  Often, the "-test" suffix indicates that it's a
testing branch used to sanity check a given branch, and if it succeeds
it will be pushed to the "normal" branch with the same name, minus the
"-test" suffix.  Dropping the "-test" suffix when comparing means the
build does not neet be triggered if tip SHA1 is the same.

Change-Id: I096196b596aa2e476f2a6feed2043ebc9aa9a162
Signed-off-by: Kevin Hilman <khilman@linaro.org>