tcwg-abe-test*,tcwg-buildfarm-test-builder*: Cleanup jobs and their names
For historical reasons tcwg-abe-test job was picked as name for
gerrit-triggered job, while tcwg-abe-test-manual as name for web-ui-triggered
job.  Similarly for tcwg-buildfarm-test-builder[-manual] jobs.
This is confusing in the whole scheme of job names, so we are cleaning this up.
The cleanup involves removing and re-creating jobs in jenkins.

Furthermore, gerrit trigger for tcwg-buildfarm-test-builder proved to be
useless, so we are removing that job.

Change-Id: Ic4ddb6d185375fa234641ee964a5f7e35d7d88a8
Signed-off-by: Maxim Kuvyrkov <maxim.kuvyrkov@linaro.org>