lt-qcom-ubuntu-images-arm64: generate rootfs specific hwpack
so far, we only generate 1 hwpack that we install in all images we produce for a
given board. Since we need graphics, multimedia, .. packages for desktop images
we put them in the hwpack, and as such they also end up (with all their
dependencies) in the non desktop images (developer, nano, ...), and that creates
sub optimal dev images.

so, we now create the hwpack for each rootfs/image that we build, so that we get
a chance to customize it for each image. We try to keep most of the hwpack
common between all images by using the .default config file.

Change-Id: I0e00f315c4d917352129c59c7dd880264827914f
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>