tcwg-backport-gerrit: Use $log_server as the coordinator node
Even though the backport job does nothing itself and only waits for other
jobs to succeed, it still requires an executor on a host.  This is due to
limitation in jenkins that a given job can't be marked as fly-weight,
similar to matrix master jobs.  If we allow "coordinator" backport jobs on
real builders, that (1) will waste builder time (the job will be idle 99%
of the time) and (2) can deadlock builders if a bunch of backport coordinator
jobs will fill up all available executors, leaving nothing for real builds.

Specifying a single node makes that node a single point of failure, but
log_server is such a point by design anyway.  Therefore it is OK to bind
backport coordinator jobs to the log_server.

Change-Id: I8f1a00b09ed8b8431f64d599b4ee40789c446ded
Signed-off-by: Maxim Kuvyrkov <maxim.kuvyrkov@linaro.org>