lt-qcom-openembedded-images: publish images on 96boards or linaro.org as appropriate
The QCOM OpenEmbedded build produces images for IFC6410 and DragonBoard 410c. We
want the IFC6410 artifacts to be published on linaro.org, while the DB410c files
need to go on 96boards.org. So we need to split the 'publishing' parts into 2
separate jobs, now that the Jenkins publishing plugin lets us do that. And we
can get rid of the previous hack (lt-qcom-openembedded-images-publish.yaml)
which never even worked..

Since this is a matrix job, i have added 2 'publish' shell scripts, for each
target. Based on the name of the machine, we publish to 96boards.org or
linaro.org. We will eventually add more machines/tests if needed.

Change-Id: I42d08de0aee6cc8e48c0c0fdc507eaa32162a862
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>