tcwg-benchmark: align parameters, properties and template.
Parameters given to the job are all used to set properties that are
meaningful for the job template. post-build-lava receives only the
settings that it needs. pbl parameters and substitutions for jobdef
are all in caps, other parameters (locals, in some sense) are in
lower-case. Add TODOs to points that will need to change when
uinstance goes live.

Change-Id: Iea81a741e3eb7af58554cbc3949a4e5cca69e19f