tcwg-benchmark: Clone via shell, getting a subdir
The main reason for this change is to clone abe into a subdir, so
that the two repositories that we clone end up in separate trees.
But as it seems silly to use Jenkins for one clone and shell for
the other, we also switch to use shell for both.

Change-Id: I1b4265984ff6a88a9fad219654fc82f28c470e25