tcwg-benchmark.sh: Export vars by line, globally
Previous approach does not work when variable values contain
spaces, so we force reading line-by-line and enquote what we set.

As we are no longer running Benchmark.sh in a subshell, we don't
need special code to try to catch its errors.

However, we are now unable to limit the parameters available to
yaml-to-python.

Change-Id: Iefbab021b640f556df55dbdc1443713371f581d7