tcwg-buildfarm-2steps-master-cambridge tcwg-buildfarm-master-cambridge tcwg-buildfarm-master-hetzner tcwg-buildfarm-master tcwg-buildfarm-maxim: Trap EXIT only
Trap will run when script exits, including in all the cases we
explicitly trapped before. But now it will only run once, even
if we have an abnormal exit.
In case of abnormal exit, the exit code is 128+signal number.

We also remove the use of $result in trap.

tcwg-buildfarm-master-hetzner-bernie: Force initialization of
$result for consistency.

Change-Id: I3843317036e8dbc426a8498dced6f9a36d608198