lt-qcom-debian-images-arm64: use mkfs.ext4 instead of make_ext4fs
Various issues have been reported about make_ext4fs, especially when it comes to
do online resizing of the rootfs. It is much preferred to use mkfs.ext4
"standard" tools to manipulate Linux root file systems.

The main advantage of make_ext4fs was to create sparse images that can be
flashed with fastboot, however along with make_ext4fs tools, there is a
standalone utility that can sparse ext images (ext2simg).

So we are replacing make_ext4fs by mkfs.ext4 + ext2simg in this patch.

We also reduce the size of the file system to a minimal size, since our root
file system includes an init script that will resize the rootfs on the first
boot to fit the entire partition.

Change-Id: Ib71cd52a4988599726e8772ea779e2789db05b1b
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>