tcwg-buildfarm-master-hetzner-bernie: Change targets
Add builds for the two targets we want to benchmark.
Remove build for native. Not looking at benchmarking that case yet,
was just doing so to work around lack of x86 builders in uinstance.

Don't enable other targets, just to avoid wasting the resource on
building things that this job isn't here to test.

Change-Id: Id274f90433b1b2954272ab1c7e089b5bf9eef348