96boards-reference-platform-openembedded: append build number to all output files
So far we only add BUILD_NUMBER to the root fs image, however in these builds we
are also deploying the kernel image file, and in some cases (for DragonBoard) we
are creating the DT and boot images. So, let's append BUILD_NUMBER for
consistency on these files too.

Also, until now , we were overriding the file auto.conf completely, which is a
bug, so instead we now append to the existing file.

Change-Id: Icdf2ea54d4e2135806007e50f13b638d7e84fa26
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>