96boards-Reference-Platform-Openembedded: add support for 'build' folder
The OE RPB setup scripts now support a dedicated build folder, so that the same
workspace can be used for multiple builds. The build folder can be provided as
argument when calling setup-environment, so let's do it.

also add some useful debug info, and display in the build log the content of
generated config files.

Note that this change is backward compatible, so it would still work with the
old setup scripts.

Change-Id: Id74b9146fef7c57899530323823f1a91ece8e55f
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>