tcwg-bkk16-buildfarm-benchmark: Eliminate race
We were triggering LAVA jobs before publishing the artifacts,
for reasons spelled out in the long comment that this patch
removes. Now that LAVA jobs are building more efficiently (see
abe commit cfac0279e002c1e1fdd2e95118ed50b84f5a8a76) it is
likely that LAVA jobs will try to access unpublished artifacts.

Now we use trigger-parameterized-builds after publishing the
artifacts. This happens on the master of the matrix job, not in
the slaves, so we have to copy the bench_parameters file from
slave to master.

Because there is no matrix-intelligence in
trigger-paramterized-builds, we have to spell out an explicit
trigger with the full path to the bench_parameters file for each
job that we want to submit. This creates 2 dependencies with
other parts of the job config:
1) If we change the list of targets for which we benchmark in the
   shell stanza, we must also change the
   trigger-parameterized-builds stanza accordingly.
2) If we change the axes of the multinode job, we must also
   change the trigger-paramaterized-builds stanza accordingly.
Comments pointing to the trigger-parameterized-builds stanza are
added to both afffected locations.

We might have a bench_parameters file from an old run hanging
around in the master workspace. Therefore it is no longer
sufficient to create bench_parameters conditionally in the slave.
Instead, we make it empty in the slave, then conditionally give
it content. If the slave does not create bench_parameters content
then the empty file gets copied to master and does not trigger a
build.

When ci.linaro.org uses a JJB with a fix for
https://storyboard.openstack.org/#!/story/2000444 in it, we can
use files from matrix slaves in trigger-parameterized-builds and
get rid of all this complexity.

Change-Id: Ife12844725b4916b88286e8aca4922ac9a5ce625