khilman-trigger-flex: add ACPI build for arm64
There is currently zero testing coverage for the ACPI support in the
upstream arm64 kernel. So as a first step, add this configuration to
the list of configurations that gets built for each tree that we build
the arm64 kernel from.

Note that CONFIG_ACPI currently depends on CONFIG_EXPERT, and defaults
to 'y' if the latter is set, so the CONFIG_ACPI=y is technically redundant.
However, it does make the description self-documenting, and it will keep
working once we remove this dependency.

Change-Id: I27e5e357389f16c1d023336efd0e255dc974bfcf
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>