khilman-trigger-flex: build Thumb2 multi_v7_defconfig with PLTs
When building the kernel in Thumb2 mode, the range of a relative jump or
call instruction is only +/- 16 MB, which may be insufficient for a module
loaded at the base of the module region to reach the entire kernel.

Since multi_v7_defconfig produces a fairly big image, actual examples can
be found on kernelci.org (e.g., socfpga) where modules fail to load due to
this issue when booting the multi_v7_defconfig/Thumb2 builds.

Since the code to deal with this is already in place
(CONFIG_ARM_MODULE_PLTS), enable it for this particular configuration.

Change-Id: Id56fe318e85edb221556859062efb7fa505be4a9
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>