art: tip: Refresh microbenchmarks presubmit.
For some unknown reason, the job does not appear at all on
CI.l.o, probably because of solar radiation, gravitational waves,
blocked chakras, or an electromagnetic monster roaming in the
servers room.

Whatever the reason, an update is in order to give this fine job a
second chance -- we all deserve a second chance at being XMLified.

Change-Id: Ia3eee9310f8274c89c93d4de28c22ed0f8c18c19
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>