art: vixl: Make way for SSH reconfiguration.
In order to clone the private repositories, the PRIVATE_KEY
is needed. For it to be available, some reconfiguration must
happen to SSH, namely the acknowledging of remote hosts' keys
and the user and host parameters in ~/.ssh/config.

In some Docker instances, the ~/.ssh directory is mounted from
the host it is running on. This means that the shell, no matter
how many sudo's it runs, will be unable to change some files or
permissions there.

This simple work-around renames the ~/.ssh directory, so that
the scripts start afresh and without any obstacles.

Change-Id: Id139926af3b73721a0c92847e2bc606592a8c898
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>