art: vixl: Install OpenJDK 7, apply patch for Stable.
Building Marshmallow requires OpenJDK 7, which is not
readily available on Ubuntu Xenial. This patch adds a
repository of OpenJDK with the most up-to-date version for
Xenial. Very cool.

A patch needed for recent versions of Clang is not yet
applied on the lor_marshmallow branch, meaning that if we
really need to build it, we need to apply this patch
somewhere somehow. There are several options at hand, but
applying the patch here seemed to be the best choice for
the time being.

Change-Id: I606776948f31651b28a481390b9da3400a616abb
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>