lt-qcom-debian-images-tracking-arm64: include all arm64 DTB in the boot image
For the tracking branch, let's not worry about which platform/DTB file to
include, and get them all. The build produces a single boot image that includes
the Image file as well as a table with all DTB, the bootloader will figure out
which DTB to load at run time.

Change-Id: I808bf5b4fc4dbd203603357b698a5c05b8426045
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>