art: stable: Run VIXL tests on Utopic.
Instead of running the VIXL tests on Ubuntu Xenial, a far
newer base system, use Ubuntu Utopic, which is a more
contemporary system of the Stable (Marshmallow) branch.

This makes away with the need to download OpenJDK 7, as it
is provided by the system, but does require
clang-format-3.6, which is unavailable; the good news is
that Ubuntu Trusty's clang-format-3.6 does work on this
version of Ubuntu (Utopic) as well.

Change-Id: I10dfd0cefe570dc54ebc9c5b1afc51242511336a
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>