tcwg-make-release: Remove useless canadian cross build.
We need only build 3 toolchains per target: x86_64, i386 and mingw.

So, perform a canadian cross build only when building on
a x86_64 slave.
We used to build twice the mingw toolchain, once on a i386 slave,
once a on x86_64 slave, risking a race condition when copying the
tarball to the fileserver.

Change-Id: Ie7ad5a69270be0b50bf9fdb2782b1217a68d0cf4