lt-qcom-debian-*: fixup the cleanup_exit trap
When the job fails , we might be in the middle of the image creation, in which
case we have a bunch of files which were extracted with root, which can't be
deleted by cleanup_exit(), which was run as regular user. Let's give super
powers to the cleanup_exit() trap.

Change-Id: I7444952b8cecf65441529d16bad4e5c61ebf5442
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>