lt-qcom-debian-images-arm64: split sd card installer into its own shell script
So that we can later reuse the builders.sh script for builds when there is no
need to create an SD card. The SD card installer script is board specific (at
least for now), so use a board specific name for the script.

Put the publishing script snippet back into the .yaml file, as this part is very
job specific, and not generic.

This commit should be a no-op.

Change-Id: Ifd4fcaa18e41a9e1489e3e745441977c180dc88c
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>