lt-qcom-debian-images-arm64-dragonboard820c: add new build job
This change adds a new job for the Debian build for Dragonboard 820c. At this
point this is used as a development vehicule until the board is finally
launched. We will not support an LTS kernel for this board until things settle
down upstream. So the kernel version will be upgraded from time to time.

Change-Id: I73020d528e2db4f18ff735d2878742cb43afe370
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>