lt-qcom-debian-images-arm64: fix cleanup scripts
In the previous rework of builders.sh , it was split into multiple subscripts,
so that we could reuse them. However the way the cleanup traps were done at the
end of the first subscripts, the entire workspace was wiped out, so the next
subscript failed. So let's defer the workspace cleanup until the last piece
(publishing), and only do early cleanup in case of failure.

Change-Id: I28de390a7a3e4768649680e974dc6456dab037e5
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>