Inject Gerrit variables before anything fails.
Building upon d11f7fb, it was noticed that a build step can
fail and Jenkins will not execute the rest of the build
steps, heading straight into the post-build steps.

Since during the build a fatal error may occur, the Gerrit
variables may not get defined at all before reaching the
email-ext plug-in.

This commit moves the recipient list to the beginning of
the job definition, so that it gets executed regardless of
the build result.

Change-Id: I8fe03cd5cbc7eebc176e151b71b890387bc26818
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>