jdk: Remove empty parameters
The empty parameters in jdk8-publish-and-report were overriding the
values passed into jdk8-publish-all-results. The parameters should be
present in each multi-job, up to the highest level job. Parameters not
present in an intermediate level are not passed down. The highest level
job determines the values of the parameter at all lower levels.

Change-Id: I5e77cd82b76a16ac8f7ea17c3fd8a008378eff6d