tcwg-buildfarm: Update from tcwg-buildfarm-maxim
This is a major update to switch buildfarm to docker.
Builds are done in docker containers.  Cross-tests are done
on test-armv8-01.tcwglab for gcc-5 (and, temporarily,
aarch64-linux-gnu gcc-6), and inside container on test-armv8-02.tcwglab
for aarch32 linux toolchains.

The updates to tcwg-backport-multijob.yaml and
tcwg-bkk16-trigger-extended.yaml jobs preserve intended semantics of
build matrix configuration.

Change-Id: I9bd024ac988dc883ce80e637fb124fcc3c04c8b4