android-lcr: make common builders.sh friendlier to set -xe
This change allows LCR builds to run with bash -xe, as it
avoids printing out a private key to the logs, as well as
error proofing an action that is bound to fail on certain
Docker slaves.

Not that we need to run with set -x, but this may help to
prevent future accidents.

Change-Id: I12595ead1e669aded60b4a4f6ec7ab4661e916df
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>