rpb-openembedded: move publishing into a separate shell step
On a fresh/clean build, the build time can exceed the publish token
timeout and will cause a build failure.
Move the linaro-publish-token closer to the actual publishing, and avoid
to take the build time into account.
To be able to publish in a separate step, we also need to inject
DEPLOY_DIR_IMAGE and MACHINE0 environment variables so it is propagated to
the other steps.

Change-Id: I2c83549859138d0f1a18b8ed8391e03e955ce806
Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>