rpb-openembedded: get rid of SPLITMACHINES and MACHINE0
We have only one platform with a secondary image and it's considered a
hack. As we don't expect to extend it to other platforms and eventually
have a multilib build, get rid of this mechanic to split machine value.
We keep hikey-32 as a special case, and it's still easier to extend later
on. As a result, it improves the readibility of the job.

Change-Id: I0212b2dff7146c5e1964ffefea5cb530c1997f8a
Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>