jdk: Pass YEAR and DAY_OF_YEAR envars
There is an issue where jobs aren't able to run across different days,
and can't be restarted across different days.
This change causes jdk8-update-src-tree.yaml to generate the YEAR and
DAY_OF_YEAR envars, which it puts into a property file the job
subsequenetly uses as parameters to launch the other jobs.

Change-Id: I70ef3cf00deb0c248d99ad20ba8bb7a49b1bc032