leg-cloud-image: use docker xenial arm64 slave
Fix "ERROR Host does not support any virtualization options"
caused by a version mismatch between libvirt daemon and client.

The libvirt client 1.2.x in CentOS 7 is too old to work with Xenial host
running libvirt daemon 1.3.1.

Adjust to run the same libvirt version for the daemon and client.
Include also a call to virt-host-validate since it's output is useful in
the build log.

Change-Id: I2e8ef5d160d2b9cebe6ee560f7b92a2821f42f90
Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>