x15 M-LCR premerge ci build: only submit single node jobs
Consider there are only 3 X15 boards on LAVA, and we are focusing
more on the boottime optimizations at the moment, so keep only
submit boottime and single node jobs only for the premerge ci build.

Regarding to the benchmark tests, we still have them submitted for
hikey and juno platforms, so we still could check if there is any
benchmark regression with them.

And we have benchmark jobs submitted when the changes merged,
so we still could confirm if there is any performance regression
on X15 platform

Change-Id: I12c17e00203db5d62cec7945ecf6aa3c2c429689
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>