lt-qcom-db410c-bootloader: add support for signing LK ELF file
on 410 platforms we are going to move to a newer SBL firmware that will require
signing the LK binaries. This feature had been disabled so far in the SBL builds
we got from QDN, but it needs to be re-enabled for commercial customers that
might need the secure boot.

The signing is a 'fake' one for community boards, and a set of scripts have been
released (open source) to make the 'fake' signing.

Change-Id: Ic9768a293d0911e9f899abf34257837788f11bb2
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>