lt-qcom-debian-images-*-arm64: add a parameter for boot image pagesize
on DB820c the boot image must be aligned to 4096, not 2048, this is mostly a
bootloader requirement. So make a parameter in the generic build script and let
the main job/yaml control the page size.

For DB410c, we keep 2048. For the -tracking job we set it to 4096 since we aim
to be able to run this job on any arm64 dragonboard.

Change-Id: Ied1c408f01e8efaf1a5ac7a433514ca18fdcf02b
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>