lt-qcom-debian: remove wcnss-bt and wcnss-wlan packages
The packages are no longer needed for a functional WLAN/BT. In fact they have
been deprecated a long time ago (since we reworked the wcn driver in the 4.4
branch around June 2016). They were kept for backward compatiblity with older
kernel, but it's 2017 soon, so let's think about the future and forget the past!

Change-Id: I8fb4a76e9bfd87d6e944933fa7addebc669dc6bb
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>