lt-qcom-debian-images-arm64-dragonboard820c: update kernel branch name
We mistakenly used release/qcomlt-4.9 name for the db820c branch, but this
wasn't really a 'release' branch, only used for engineering drops/builds. Now
that we really need to support the 4.9 LTS release, we need to rename this
branch. What used to be 'release/qcomlt-4.9' has been pushed in the repo as
'db820c/qcomlt-4.9' , and it is now clear that this branch is only used for the
DB820c bringup.

Change-Id: I0ec084564a868547fcbc080bf5e10eeb791dc7fb
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>