tcwg-llvm-build.yaml: Adding new toolchain parameter
This new parameter allows us to choose a toolchain to use to build LLVM.
This can either be an upstream release (from llvm.org/releases) or
internal ones stored in dev-01, so we can do stage 2 builds.

A similar patch in jenkins-scripts is being submitted to cope with the
new parameter.

Change-Id: I22a66c324fbe5320b9a6d1eb4710678f090a2ef6