tcwg-buildfarm: Call start-container-docker.sh with --no-ssh-agent.
Since Jenkins' ssh-agent plugin was updated, we hope that this
workaround is not needed anymore. Try with this option before
actually removing the workaround from start-container-docker.sh
itself.

The expected impact is on tk1/tx1 slaves.

Other jobs could use the same flag, but let's try with this one
first.

Change-Id: I48a0b2d3cf37e5ea70d76dcc51191befec43b601