tcwg-llvm-release.yaml: Chain test-suite job after release
This is an attempt to chain a test-suite job, like the staged-build,
after the release. IIUC, the chaining will not happen if the container
script fails (any test failure), which is *not* what I want.

I'm looking for ideas on how to do that. If we need to use the same
trick as the release.sh script, I'll update the patch to do that.

Change-Id: I1ab593c27edcbe118bffaafa9c2498c382fade4b