tcwg-*: Switch all matrix and multijob jobs to tcwg-coordinator
This change makes all TCWG matrix and multijob "coordination" jobs
to run on tcwg-coordinator label.  The "coordination" jobs do
not consume significant resources and just wait for other jobs
to finish.  In case of matrix jobs these are "flyweight" jobs,
and all TCWG multijobs behave similarly.

Another benefit from this change is that git checkouts
for "coordination" jobs will no longer be mixed with workspaces of
matrix sub-builds of the same jobs, if sub-builds happen to start
on the same node as coordination job.

This has been auto-generated with the following script

$ cat > tcwg-coordinator.sh <<EOF

for f in tcwg-*.yaml; do
    type=$(grep "project-type" $f | sed -e "s/.*: //")

    case $f in
	tcwg-bkk16-buildfarm-benchmark.yaml) continue ;;
    esac

    case $type in
	freestyle) continue ;;
	matrix|multijob)
	    sed -e "s/ node:.*/ node: tcwg-coordinator/" -i $f
	;;
	*) echo "ERROR: Wrong job type: $type"; exit 1 ;;
    esac
done
EOF

Change-Id: I216abbf903537da2c2bfd3f6d5a1fbf6b8fe4877