tcwg-abe-test, tcwg-validate-abetests: Fix artifacts.
Clean up reporting and artifact creation for abe-tests.

This patch makes the reporting and artifacts consistent between
the tcwg-abe-test job and the tcwg-validate-abetests job. This
fixes overly verbose reporting for tcwg-abe-test, and ensures that
tcwg-validate-abetests produces artifacts when the test fails, as
previously "set -e" prevented the copy of the logs to ${ARTIFACTS}.

This should also fix "xz: Cannot set the file group: Operation not
permitted" errors when the artifacts are compressed, which result
from inconsistent group membership between the host and docker
containers.

Change-Id: Ib47a80db81cfc2baf03b65aa4b83f552b660e325