tcwg-lvm-*: Add compiler option to the builds
This change adds a "compiler" option to override the system compiler if
no toolchain is provided by the parameters. The default is clang because
that's known to work on all architectures, while GCC 5.4 (Xenial) is
known to fail on ARM32 and ARM64 in some configurations.

The scripts need to be changed accordingly to set CC/CXX on their own
based in the new argument.

Change-Id: I0632d3a7d2c53650ac74ba5f2576774c6c7f7cf1