optee-build: set child-workspace
EDK2 has a restriction in the length of the path.
When we use matrix job, the axis are part of the path and increase the
length of the path. It's causing EDK2 to fail to build.
Set the child-workspace to "." in order to workaround this EDK2 limitation
and limit path length.

Change-Id: I08e507df9fd433ff3065980f6297c0ac7a8d42e7
Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>