tcwg-llvm-release.yaml: Move Release job to use llvmprod branch
This is our first attempt at stabilising the production environment in
Jenkins. For now, we're only fixing jenkins-script, but in the future,
we'll need to make sure the dockerfiles and the LLVM scripts are all
locked up in the same production branch.

By making it into a parameter, we also allow for master/staging
validation, as well as snapshot production+patch to cherry pick fixes if
needed.

Change-Id: Ida36dc7dab510a41227fb793be8e5fa0c5ed8fb0