tcwg-llvm-precommit.yaml: New job
This is the first attempt at a pre-commit job. For now, we're only
building the branch, on all targets at once, with limited options.

The next step is to find the last revision of all projects in that
branch and start a second build, on master, at that revision, as a
baseline.

Needs https://review.linaro.org/19818

Change-Id: I0edb3d24ce51a8dfb3017862276755b21593e793