tcwg-compare-results: Add pass_thresh parameter. tcwg-trigger-binutils: Set pass_thresh to 0.92.
Some binutils confguration have a low pass rate (about 0.92), because
of many untested/unsupported tests.  This causes the non-regression
comparison to return an error even if there was no regression, because
the default minimum expected pass rate is 0.95.

Use the new -pass-thresh comparison parameter, to override the value
for binutils and set it to 0.92 to avoid reporting errors all the
time.

Change-Id: I0174079b73fb56a3ed3252b04329c50b93f52e7b