lt-qcom-debian-images: remove this job from triggers
We want to eventually remove this job and transition to
lt-qcom-debian-images-dragonboard600c instead, as the reference job. It is not
very much maintained right now, but we still need to keep it around for now. So
let's remove it from the various triggers at least. It can be manually started
when/if needed.

Change-Id: Iddd076b6dd8dac3386d63e8a78c02d487607bce3
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>