zephyr-upstream, zephyr-aeolus: libc6-dev-i386 package is no longer needed.
It was needed for older versions of Zephyr SDK which were 32-bit, but for
few last releases, the SDK was 64-bit. In the meantime, libc6-dev-i386
is absent in the Docker image, and its install during build causes also
upgrade to bunch of libc-related packages, taking 20-30 mins of build time
(half of it in total).

Change-Id: I37054287c239fdd446a671516a215576cc80f15c
Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>