openembedded-lkft: promote KERNEL_DESCRIBE and SRCREV_kernel to job parameters
The build triggers have already all the git describe and commit for each
kernel. Instead of computing them in the build job, pass them from the
trigger job to the downstream job. As a result, KERNEL_DESCRIBE and
SRCREV_kernel are now job parameters.

As a side effect, it allows to simplify build-name-setter and directly use
SRCREV_kernel variable.

Change-Id: I01c823259fa98484cd27ef163a66327a694262a1
Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>