tcwg-llvm-testsuite: Add 'nt_flags' field
Add a 'nt_flags' field to the job which will be forwarded to the script
that builds the test-suite. This should enable us to build the
test-suite with different compiler flags (e.g. -O0 -mllvm -global-isel),
or in benchmarking mode, or with a specified number of threads, or
anything else that the LNT nt producer allows us to configure.

Change-Id: Iaa78f499d55ee90fa0b9eb1f136c5cf816dc72e1