tcwg-llvm-*: Add 'scripts_branch' field
We used to have the scripts_branch field only on the tcwg-llvm-release
job, which meant we could only stage the scripts used by that job. This
commit adds scripts_branch to all the other LLVM jobs, so we can
properly stage them when needed.

The default value of the scripts_branch is llvmprod, which is the
default we had for the release job. This is different from master, which
is the value we previously used in these jobs. The new default is a 
better choice, since that is the stable branch meant to be used in
production.

Change-Id: I6c64eca5e1bf87163edbc1e624f5b5a28d770321