tcwg-update-snapshots_ref: Fix u-s_r for all Jenkins nodes.
As jenkins-helpers.sh was not sourced in the script,
print_nodes_in_labels was not defined.

Also, because the reference to print_nodes_in_labels was in the
start of a for loop, the error was not caught by 'set -e'. To
avoid this in future, move the call to print_nodes_in_labels
into a separate array assignment.

Change-Id: Ib46dd81174aeeb1da03551770c7763c8c4191db9