linaro-hikey-stable-rc-4.4-merge-and-tag: try to merge from new base
For the merge, use Sumit's lts-4.4.y-hikey, which gets
updated every time a 4.4.y stable kernel is released. This
is a good base branch since it is Guodong's original branch
rebased with the latest 4.4.y kernel.

If a conflict arises from the merge operation, then abort
the whole thing; otherwise, let the process continue as
expected.

A merge failure here indicates an important conflict in the
code that needs to be resolved manually by a human (a
knowledgeable one) so that the series of successive merges
do not follow a wrong path.

Change-Id: Ib699fccbd7ac5bc0046a50aaf7a6c89db23f1b7c
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>