tcwg-llvm-testsuite: Fix nt_flags default
We used to get unbound variable errors when the nt_flags were not set.
This is especially problematic because the release job launches the
test-suite with the default, unset nt_flags, which made all release jobs
fail.

The fix is to pass --nt-flags to the test-suite script when the Jenkins
nt_flags parameter is set and nothing when it is unset. Thanks to
Charles Baylis for the exact bash incantation for achieving this.

Change-Id: I41786f30ab6394bec8e8f9147bc8e93ea910940e