trigger-openembedded-lkft-linaro-hikey-stable-rc-4.4: better commit description
Switch to git cli to avoid to rely on Jenkins' Git SCM plug-in.

The Git object database did not include enough references to
provide a good description of the commit that triggered the
pipeline. Adding the linux-stable-4.4 git tree helps a lot
in that references (KERNEL_DESCRIBE) are now based on stable
tags.

Change-Id: I9afd14d5b80b5cace87cfe439abc6f18599b1d6b
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>