trigger-openembedded-lkft-linaro-hikey-stable-rc-4.4: allow building arbitrary tags
This job is triggered whenever a new change occurs in GKH's
linux-stable-rc tree, branch linux-4.4.y, via the merge-and-
tag job. If said change is difficult to merge automatically,
the merge-and-tag stage will fail and will require human
intervention.

Now, when a human correctly corrects the correction and the
new tag is pushed to the lkft/arm64-stable-rc tree, there
is no way to resume the process from where it stopped. This
trigger, being the next step after the merge-and-tag job,
should then be triggered by the human via the Build Now
button to continue with _that_ tag.

It is now possible to trigger this job (and the build there-
after) with a specific tag, through the parameter
LINUX_TAG_PATTERN.

Change-Id: Icf8014210019d1ed0c1800f05e5da26c7c8c0a37
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>