mbl-openembedded: build morty and master
Create separate jobs for morty and master.

We are releasing from morty, which is stable, but we also want to track
master to make sure nothing upstream takes us by surprise.

Extend the build matrix to build both default.xml and
pinned-manifest.xml, base on whether we are building from the pinned
branch or not.

Change-Id: I35e4598a2b1b2ea5b00a9981a8d69eb9d51517de
Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>