lt-qcom-linux: do not use BUILD_NUMBER in source package name
Let's step back and not use the BUILD_NUMBER as the package ABI. We already have
the kernel version (x.y) in the source package. If we use BUILD_NUMBER in the
source package name we are going to end up with a huge numbers of source
packages in OBS.

We can add -1, -2, if needed , when ABI breaks are detected.

Change-Id: I1eb1986e95671ded133b5e3574890d725421b7af
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>