FIXME: clean sstate-cache for aarch64_ilp32
This is causing serious issues after upgrading GLIBC, as the
ABI has changed and those changes are not picked up correctly.
Work around is to remove sstate-cache although only for those
file that have been previously been linked with GLIBC (i.e.
all that's in the cache for architecture aarch64_ilp32).

Failures look like this:
| [...]/build/tmp-glibc/work/aarch64_ilp32-oe-linux-gnuilp32/gnutls/3.5.13-r0/recipe-sysroot/usr/lib/libstdc++.so: undefined reference to `__strxfrm_l@GLIBC_2.26'
| [...]/build/tmp-glibc/work/aarch64_ilp32-oe-linux-gnuilp32/gnutls/3.5.13-r0/recipe-sysroot/usr/lib/libstdc++.so: undefined reference to `__nl_langinfo_l@GLIBC_2.26'
| [...]/build/tmp-glibc/work/aarch64_ilp32-oe-linux-gnuilp32/gnutls/3.5.13-r0/recipe-sysroot/usr/lib/libstdc++.so: undefined reference to `aligned_alloc@GLIBC_2.26'
| [...]/build/tmp-glibc/work/aarch64_ilp32-oe-linux-gnuilp32/gnutls/3.5.13-r0/recipe-sysroot/usr/lib/libstdc++.so: undefined reference to `wmemcpy@GLIBC_2.26'
| [...]/build/tmp-glibc/work/aarch64_ilp32-oe-linux-gnuilp32/gnutls/3.5.13-r0/recipe-sysroot/usr/lib/libstdc++.so: undefined reference to `strcmp@GLIBC_2.26'
and typically involve glibc, gnutls, and libcgroups.

Cleaning up those packages is not enough, hence the general
clean-up.

This ought to be removed once GLIBC upgrade is stable.

Change-Id: Iece70f4f9a6407301fa5d7d67b51a61ecc18d94b
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>